TUP-5221 fix: shouldn't do update when logon local mode project https://jira.talendforge.org/browse/TUP-5221
