+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/ui/login/LoginProjectPage.java
@@ -87,7 +87,6 @@ import org.talend.core.model.general.Project;
 import org.talend.core.model.properties.ProjectReference;
 import org.talend.core.model.properties.User;
 import org.talend.core.model.repository.SVNConstant;
-import org.talend.core.prefs.PreferenceManipulator;
 import org.talend.core.repository.model.IRepositoryFactory;
 import org.talend.core.repository.model.ProxyRepositoryFactory;
 import org.talend.core.repository.model.RepositoryFactoryProvider;
@@ -995,13 +994,6 @@ public class LoginProjectPage extends AbstractLoginActionPage {
         if (LoginHelper.isRestart) {
             loginDialog.okPressed();
         } else {
-            // This is for check if the reference project need to update
-            try {
-                saveUpdateStatus(getProject());
-            } catch (Exception e) {
-                // show exception
-                MessageBoxExceptionHandler.process(e, loginHelper.getUsableShell());
-            }
             boolean isLogInOk = loginHelper.logIn(getConnection(), getProject());
             if (isLogInOk) {
                 LoginHelper.setAlwaysAskAtStartup(alwaysAsk.getSelection());
@@ -1017,57 +1009,6 @@ public class LoginProjectPage extends AbstractLoginActionPage {
         Context ctx = CoreRuntimePlugin.getInstance().getContext();
         RepositoryContext repositoryContext = (RepositoryContext) ctx.getProperty(Context.REPOSITORY_CONTEXT_KEY);
         repositoryContext.setNoUpdateWhenLogon(false);
-    }
-
-    private void saveUpdateStatus(Project project) throws JSONException {
-        Context ctx = CoreRuntimePlugin.getInstance().getContext();
-        RepositoryContext repositoryContext = (RepositoryContext) ctx.getProperty(Context.REPOSITORY_CONTEXT_KEY);
-        PreferenceManipulator prefManipulator = new PreferenceManipulator();
-        if (!LoginHelper.isRemoteConnection(getConnection())) {
-            repositoryContext.setNoUpdateWhenLogon(true);
-            return;
-        }
-        String url = project.getEmfProject().getUrl();
-        if (url == null || !"git".equals(getStorage(url))) {
-            return;
-        }
-        String location = getLocation(url);
-        String projectName = project.getTechnicalLabel();
-        String branch = getBranch();
-        if (branch.startsWith("tags/")) {
-            repositoryContext.setNoUpdateWhenLogon(true);
-            return;
-        }
-        if (branch.startsWith("branches/"))
-            branch=branch.substring("branches/".length());
-        JSONObject json = prefManipulator.getLogonLocalBranchStatus(location, projectName);
-        if (json != null) {
-            if (!json.has(branch))
-                return;
-            Object noUpdateWhenLogon = json.get(branch);
-            if (noUpdateWhenLogon != null) {
-                repositoryContext.setNoUpdateWhenLogon((Boolean) noUpdateWhenLogon);
-            }
-        }
-
-    }
-
-    public static String getStorage(String url) throws JSONException {
-        JSONObject jsonObject = new JSONObject(url);
-        String location = ""; //$NON-NLS-1$
-        if (jsonObject.has("storage")) {
-            location = jsonObject.getString("storage"); //$NON-NLS-1$
-        }
-        return location;
-    }
-
-    public static String getLocation(String url) throws JSONException {
-        JSONObject jsonObject = new JSONObject(url);
-        String location = ""; //$NON-NLS-1$
-        if (jsonObject.has("location")) {
-            location = jsonObject.getString("location"); //$NON-NLS-1$
-        }
-        return location;
     }
 
     @Override
